refactor(observability): 分离体验会话故事
